@@ -15,7 +15,7 @@
  */
 
 plugins {
-    id 'com.github.rodm.teamcity-agent' version "1.0"
+    id 'com.github.rodm.teamcity-agent' version "1.4.1"
 }
 
 teamcity {
@@ -26,19 +26,20 @@ teamcity {
 }
 
 dependencies {
-    compile project(':google-storage-common')
-    compile "org.jetbrains.kotlin:kotlin-stdlib"
-    compile "org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlinxVersion"
-    compile "com.google.cloud:google-cloud-storage:$googleSdkVersion"
-    testCompile 'org.testng:testng:6.8'
-}
+    implementation project(':google-storage-common')
+    implementation "org.jetbrains.kotlin:kotlin-stdlib"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlinxVersion"
+    implementation ("org.apache.tika:tika-core:2.4.1") {
+        exclude group: 'org.slf4j', module: 'slf4j-api'
+    }
 
-kotlin { experimental { coroutines 'enable' } }
+    testImplementation 'org.testng:testng:6.8'
+}
 
 agentPlugin.version = null
 agentPlugin.baseName = projectIds.artifact
 
 compileKotlin {
-    sourceCompatibility = JavaVersion.VERSION_1_7
-    targetCompatibility = JavaVersion.VERSION_1_7
+    sourceCompatibility = JavaVersion.VERSION_1_8
+    targetCompatibility = JavaVersion.VERSION_1_8
 }
