@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2021 JetBrains s.r.o.
+ * Copyright 2000-2022 JetBrains s.r.o.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,34 +20,58 @@ import com.google.cloud.storage.BlobInfo
 import com.google.cloud.storage.HttpMethod
 import com.google.cloud.storage.Storage
 import com.google.common.cache.CacheBuilder
+import com.intellij.openapi.diagnostic.Logger
 import jetbrains.buildServer.serverSide.TeamCityProperties
 import jetbrains.buildServer.serverSide.artifacts.google.GoogleConstants
 import jetbrains.buildServer.serverSide.artifacts.google.GoogleUtils
+import java.util.*
 import java.util.concurrent.TimeUnit
 
 class GoogleSignedUrlProviderImpl : GoogleSignedUrlProvider {
     private val myLinksCache = CacheBuilder.newBuilder()
-            .expireAfterWrite(urlLifetimeSec.toLong(), TimeUnit.SECONDS)
-            .maximumSize(200)
-            .build<String, String>()
+        .expireAfterWrite(urlLifetimeSec.toLong(), TimeUnit.SECONDS)
+        .maximumSize(200)
+        .build<String, String>()
 
     override val urlLifetimeSec: Int
-        get() = TeamCityProperties.getInteger(GoogleConstants.URL_LIFETIME_SEC, GoogleConstants.DEFAULT_URL_LIFETIME_SEC)
+        get() = TeamCityProperties.getInteger(
+            GoogleConstants.URL_LIFETIME_SEC,
+            GoogleConstants.DEFAULT_URL_LIFETIME_SEC
+        )
 
-    override fun getSignedUrl(httpMethod: HttpMethod,
-                              path: String,
-                              parameters: Map<String, String>): Pair<String, Int> {
+    override fun getSignedUrl(
+        httpMethod: HttpMethod,
+        path: String,
+        parameters: Map<String, String>
+    ): Pair<String, Int> {
         val lifeTime = urlLifetimeSec
         val resolver = {
             val bucket = GoogleUtils.getStorageBucket(parameters)
             val blobInfo = BlobInfo.newBuilder(bucket, path)
-            val urlOptions = arrayListOf<Storage.SignUrlOption>(Storage.SignUrlOption.httpMethod(httpMethod))
+            val urlOptions = arrayListOf<Storage.SignUrlOption>(
+                Storage.SignUrlOption.httpMethod(httpMethod),
+                Storage.SignUrlOption.withV4Signature()
+            )
+
             if (httpMethod.name() == "POST") {
-                urlOptions.add(Storage.SignUrlOption.withExtHeaders(mapOf("x-goog-resumable" to "start")))
+                urlOptions.add(
+                    Storage.SignUrlOption.withExtHeaders(
+                        mapOf(
+                            "x-goog-resumable" to "start",
+                            "Content-Type" to parameters["contentType"]
+                        )
+                    )
+                )
                 urlOptions.add(Storage.SignUrlOption.withContentType())
                 blobInfo.setContentType(parameters["contentType"])
             }
-            bucket.storage.signUrl(blobInfo.build(), lifeTime.toLong(), TimeUnit.SECONDS, *urlOptions.toTypedArray()).toString()
+
+            bucket.storage.signUrl(blobInfo.build(), lifeTime.toLong(), TimeUnit.SECONDS, *urlOptions.toTypedArray())
+                .toString()
+                .also {
+                    LOG.debug("signedURL: $it")
+                    LOG.debug("contentType: ${parameters["contentType"]}")
+                }
         }
 
         if (httpMethod == HttpMethod.GET && TeamCityProperties.getBoolean(GoogleConstants.SIGNED_URL_GET_CACHE_ENABLED)) {
@@ -63,6 +87,10 @@ class GoogleSignedUrlProviderImpl : GoogleSignedUrlProvider {
             append(params[GoogleConstants.PARAM_ACCESS_KEY])
             append(params[GoogleConstants.PARAM_BUCKET_NAME])
             append(path)
-        }.toString().toLowerCase().hashCode().toString()
+        }.toString().lowercase(Locale.getDefault()).hashCode().toString()
+    }
+
+    companion object {
+        private val LOG = Logger.getInstance(GoogleSignedUrlProviderImpl::class.java.name)
     }
 }
